Move all input-related classes to a new `brian2.input` package
