+++ brian2/groups/__init__.py
@@ -4,7 +4,4 @@ Package providing groups such as `NeuronGroup` or `PoissonGroup`.
 
 from .group import *
 from .neurongroup import *
-from .poissongroup import *
-from .poisson_input import *
-from .spikegeneratorgroup import *
 from .subgroup import *
