+++ brian2/input/spikegeneratorgroup.py
@@ -7,9 +7,7 @@ from brian2.core.spikesource import SpikeSource
 from brian2.units.fundamentalunits import check_units, Unit
 from brian2.units.allunits import second
 from brian2.core.variables import Variables
-from brian2.groups.group import CodeRunner
-
-from .group import Group
+from brian2.groups.group import CodeRunner, Group
 
 __all__ = ['SpikeGeneratorGroup']
