+++ brian2/input/timedarray.py
