+++ brian2/only.py
@@ -37,6 +37,7 @@ from brian2.core.namespace import *
 from brian2.groups import *
 from brian2.synapses import *
 from brian2.monitors import *
+from brian2.input import *
 from brian2.spatialneuron import *
 from brian2.devices import set_device, get_device, device
 import brian2.devices.cpp_standalone as _cpp_standalone
