+++ brian2/utils/__init__.py
@@ -3,5 +3,3 @@ Utility functions for Brian.
 '''
 
 from .logger import *
-from .timedarray import *
-from .binomial import *
